+++ packages/components/src/components/basic/marque/marque-styles.ts
@@ -1,0 +1,44 @@
+import { attachCss, buildGlobalStyles, buildHostStyles, getCss, getFocusStyles, mediaQuery } from '../../../utils';
+import { color } from '@porsche-design-system/utilities';
+import { MarqueSize } from './marque-utils';
+
+const baseSizes = {
+  small: {
+    width: 100,
+    height: 60,
+  },
+  medium: {
+    width: 120,
+    height: 72,
+  },
+};
+
+export const getComponentCss = (size: MarqueSize): string => {
+  return getCss({
+    ...buildHostStyles({
+      display: 'inline-flex',
+      verticalAlign: 'top',
+    }),
+    ...buildGlobalStyles({
+      a: {
+        display: 'block',
+        textDecoration: 'none',
+        ...getFocusStyles({ color: color.default, offset: 0 }),
+      },
+      picture: {
+        display: 'block',
+        ...baseSizes.small,
+        ...(size !== 'responsive' ? baseSizes[size] : { [mediaQuery('l')]: baseSizes.medium }),
+      },
+      img: {
+        display: 'block',
+        width: '100%',
+        height: 'auto',
+      },
+    }),
+  });
+};
+
+export const addComponentCss = (host: HTMLElement, size: MarqueSize): void => {
+  attachCss(host, getComponentCss(size));
+};