+++ packages/components/src/components/basic/typography/text/text-utils.spec.ts
@@ -1,0 +1,12 @@
+import { isSizeInherit, TextSize } from './text-utils';
+import { BreakpointCustomizable } from '../../../../utils';
+
+describe('isSizeInherit()', () => {
+  it.each<[BreakpointCustomizable<TextSize>, boolean]>([
+    ['inherit', true],
+    [{ base: 'large', l: 'inherit' }, true],
+    ['medium', false],
+  ])('should for size %s return %s', (size, result) => {
+    expect(isSizeInherit(size)).toBe(result);
+  });
+});