+++ packages/components/src/components/content/table/table-row/table-row.spec.ts
@@ -1,0 +1,17 @@
+import * as domUtils from '../../../../utils/dom';
+import { TableRow } from './table-row';
+
+describe('table-row', () => {
+  describe('connectedCallback', () => {
+    it('should call throwIfParentIsNotOfKind()', () => {
+      const spy = jest.spyOn(domUtils, 'throwIfParentIsNotOfKind');
+      const component = new TableRow();
+
+      try {
+        component.connectedCallback();
+      } catch (e) {}
+
+      expect(spy).toBeCalledWith(undefined, 'pTableBody');
+    });
+  });
+});