+++ packages/components/src/components/content/tabs/tabs-item/tabs-item.spec.ts
@@ -1,0 +1,17 @@
+import * as domUtils from '../../../../utils/dom';
+import { TabsItem } from './tabs-item';
+
+describe('tabs-item', () => {
+  describe('connectedCallback', () => {
+    it('should call throwIfParentIsNotOfKind()', () => {
+      const spy = jest.spyOn(domUtils, 'throwIfParentIsNotOfKind');
+      const component = new TabsItem();
+
+      try {
+        component.connectedCallback();
+      } catch (e) {}
+
+      expect(spy).toBeCalledWith(undefined, 'pTabs');
+    });
+  });
+});