+++ packages/components/tests/unit/components/banner-utils.spec.ts
@@ -1,5 +1,0 @@
-import { closeStyles } from '../../../src/components/feedback/banner/banner-utils';
-
-it('closeStyles should contain correct css', () => {
-  expect(closeStyles).toMatchSnapshot();
-});