+++ packages/components/CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 - `Checkbox Wrapper`, `Radio Button Wrapper`, `Select Wrapper`, `Textarea Wrapper` and `Text Field Wrapper` now reflect changes of the `required` attribute on their child component
+- `multiline` property to `Table Cell`
+
+#### Fixed
+- `Table Head Cell` uses semi bold instead of bold as font weight
+- Transition of `Modal`
 
 ### [2.3.0-beta.1] - 2021-07-08